feat: add recipe sidebar and scheduler components; refactor meal handling in recipe card fix: update meal type to Partial<Meal> for better handling of optional properties refactor: move recipe list and meal plan controls to separate components for better organization style: improve layout and styling of week view and day column components test: add unit tests for new components and update existing tests

+++ src/app/_components/grocy-recipe-card/grocy-recipe-card.component.ts
@@ -5,13 +5,13 @@ import { Recipe } from '../../interfaces/recipe.interface';
 
 
 @Component({
-    selector: 'app-grocy-recipe-card',
-    templateUrl: './grocy-recipe-card.component.html',
-    styleUrl: './grocy-recipe-card.component.scss',
-    standalone: false
+  selector: 'app-grocy-recipe-card',
+  templateUrl: './grocy-recipe-card.component.html',
+  styleUrl: './grocy-recipe-card.component.scss',
+  standalone: false
 })
 export class GrocyRecipeCardComponent {
-  @Input() meal: Meal | undefined;
+  @Input() meal: Partial<Meal> | undefined;
   @Input() recipe: Recipe | undefined;
 
   draggable = {
@@ -22,15 +22,19 @@ export class GrocyRecipeCardComponent {
 
   constructor(private grocyService: GrocyService) { }
 
-  deleteMeal(meal: Meal) {
-    this.grocyService.deleteMeal(meal.id).subscribe(() => {
-      // this.refreshMealPlan$.next();
-    });
+  deleteMeal(meal: Partial<Meal>) {
+    if (meal.id) {
+      this.grocyService.deleteMeal(meal.id).subscribe(() => {
+        // this.refreshMealPlan$.next();
+      });
+    }
   }
 
-  doneMeal(meal: Meal) {
-    meal.done = 1;
-    this.grocyService.updateMeal(meal);
+  doneMeal(meal: Partial<Meal>) {
+    if (meal.id) {
+      meal.done = 1;
+      this.grocyService.updateMeal(meal as Meal);
+    }
   }
 
   openRecipe(recipeId: number) {
